feat(curric): added tab routes and sequence layout
